services/audio-capture/requirements.txt
redis>=5.0.0
numpy>=1.24.0
soundfile>=0.12.0
pyyaml>=6.0
requests>=2.31.0
websockets>=12.0
scipy>=1.10.0
netifaces>=0.11.0
msgpack